PythonSCR/requirements.txt
//Python 3.9.0
paho-mqtt==1.5.1
pika==1.1.0
